+++ Chrominsky.Utils/Models/Base/BaseDatabaseEntityWithTenantId.cs
@@ -1,0 +1,11 @@
+using System.ComponentModel.DataAnnotations;
+using Chrominsky.Utils.Models.Base.Interfaces;
+
+namespace Chrominsky.Utils.Models.Base;
+
+public class BaseDatabaseEntityWithTenantId : BaseDatabaseEntity, IBaseDatabaseEntityWithTenantId
+{
+    /// <inheritdoc />
+    [Required]
+    public Guid? TenantId { get; set; }
+}