+++ Chrominsky.Utils/Models/Base/Interfaces/IBaseDatabaseEntityWithTenantId.cs
@@ -1,0 +1,9 @@
+namespace Chrominsky.Utils.Models.Base.Interfaces;
+
+public interface IBaseDatabaseEntityWithTenantId : IBaseDatabaseEntity
+{
+    /// <summary>
+    /// Identifier of the tenant to which the entity belongs. Nullable.
+    /// </summary>
+    Guid? TenantId { get; set; }
+}