+++ Chrominsky.Utils/Models/Base/Interfaces/IBaseSimpleEntity.cs
@@ -1,0 +1,6 @@
+namespace Chrominsky.Utils.Models.Base.Interfaces;
+
+public interface IBaseSimpleEntity
+{
+    
+}