+++ Chrominsky.Utils/Repositories/Base/IBaseDatabaseRepository.cs
@@ -1,4 +1,5 @@
 using Chrominsky.Utils.Models.Base;
+using Chrominsky.Utils.Models.Base.Interfaces;
 
 namespace Chrominsky.Utils.Repositories.Base;
 
@@ -22,7 +23,7 @@ public interface IBaseDatabaseRepository<T>
     /// <typeparam name="T">Type of entity to be added.</typeparam>
     /// <param name="entity">The entity to be added.</param>
     /// <returns>The unique identifier of the newly added entity.</returns>
-    Task<Guid> AddAsync<T>(T entity) where T : BaseDatabaseEntity;
+    Task<Guid> AddAsync<T>(T entity) where T : class, IBaseDatabaseEntity;
 
     /// <summary>
     /// Updates an existing entity in the database.
@@ -30,7 +31,7 @@ public interface IBaseDatabaseRepository<T>
     /// <typeparam name="T">Type of entity to be updated.</typeparam>
     /// <param name="entity">The updated entity.</param>
     /// <returns>The updated entity.</returns>
-    Task<T> UpdateAsync<T>(T entity) where T : BaseDatabaseEntity;
+    Task<T> UpdateAsync<T>(T entity) where T : class, IBaseDatabaseEntity;
 
     /// <summary>
     /// Deletes an entity from the database by its unique identifier.
@@ -38,5 +39,5 @@ public interface IBaseDatabaseRepository<T>
     /// <typeparam name="T">Type of entity to be deleted.</typeparam>
     /// <param name="id">Unique identifier of the entity.</param>
     /// <returns>True if the entity was successfully deleted, otherwise false.</returns>
-    Task<bool> DeleteAsync<T>(Guid id) where T : BaseDatabaseEntity;
+    Task<bool> DeleteAsync<T>(Guid id) where T : class, IBaseDatabaseEntity;
 }